requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.0
httpx==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0